Model3D、Object3D、Object3DManagerリファクタリング第一弾
・全体的な記述量削減
・パフォーマンスの1部改善
・依存関係の改善

新たに出た問題

・同じObject3dを違う場所に複数描画しようとしてもできない
（定数バッファのワールド変換行列の部分を対応する数分用意しないといけない）

+++ RVector.cpp
@@ -23,7 +23,7 @@ const RVector3 Rv3Ease::InQuad(const RVector3 &s, const RVector3 &e, const float
 {
 	RVector3 start = s;
 	RVector3 end = e;
-	return start * (1.0f - t * t) + end * pow(t, 2);
+	return start * (1.0f - t * t) + end * (float)pow(t, 2);
 }
 
 const RVector3 Rv3Ease::OutQuad(const RVector3 &s, const RVector3 &e, const float t)
@@ -37,5 +37,5 @@ const RVector3 Rv3Ease::InOutQuad(const RVector3 &s, const RVector3 &e, const fl
 {
 	RVector3 start = s;
 	RVector3 end = e;
-	return t < 0.5f ? (start * (1.0f - t)) + end * (t * t) : start * (1.0f - t) + end * (1 - pow(-2 * t + 2, 2) / 2);
+	return t < 0.5f ? (start * (1.0f - t)) + end * (t * t) : start * (1.0f - t) + end * (1 - (float)pow(-2 * t + 2, 2) / 2);
 }
